Refactored backend code as per review comments
