+++ dataproc_jupyter_plugin/commons/constants.py
@@ -12,15 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-from dataproc_jupyter_plugin.services.gcpUrlService import gcp_service_url
-
 ENVIRONMENT_API = "https://composer.googleapis.com/v1"
 TAGS = "dataproc_jupyter_plugin"
 GCS = "gs://"
 CONTENT_TYPE = "application/json"
-dataproc_url = gcp_service_url("dataproc")
-
-storage_url = gcp_service_url(
-    "storage", default_url="https://storage.googleapis.com/storage/v1/"
-)
+PACKAGE_NAME = "dataproc_jupyter_plugin"
+COMPOSER_SERVICE_NAME = "composer"
+DATAPROC_SERVICE_NAME = "dataproc"
+BIGQUERY_SERVICE_NAME = "bigquery"
+DATACATALOG_SERVICE_NAME = "datacatalog"
+CLOUDRESOURCEMANAGER_SERVICE_NAME = "cloudresourcemanager"
+STORAGE_SERVICE_NAME = "storage"
+COMPUTE_SERVICE_NAME = "compute"
+METASTORE_SERVICE_NAME = "metastore"
+CLOUDKMS_SERVICE_NAME = "cloudkms"
+COMPUTE_SERVICE_DEFAULT_URL = "https://compute.googleapis.com/compute/v1"
+STORAGE_SERVICE_DEFAULT_URL = "https://storage.googleapis.com/storage/v1/"
