+++ dataproc_jupyter_plugin/controllers/composer.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Google LLC
+# Copyright 2024 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,26 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 import json
 from dataproc_jupyter_plugin import credentials
 from jupyter_server.base.handlers import APIHandler
 import tornado
-from dataproc_jupyter_plugin.services.runtimeListService import RuntimeListService
+from dataproc_jupyter_plugin.services import composer
 
 
-class RuntimeController(APIHandler):
+class EnvironmentListController(APIHandler):
     @tornado.web.authenticated
     async def get(self):
+        """Returns names of available composer environments"""
         try:
-            page_token = self.get_argument("pageToken")
-            page_size = self.get_argument("pageSize")
-            runtime = RuntimeListService()
-            gcp_credentials = await credentials.get_cached()
-            runtime_list = runtime.list_runtime(
-                gcp_credentials, page_size, page_token, self.log
-            )
-            self.finish(json.dumps(runtime_list))
+            client = composer.Client(await credentials.get_cached(), self.log)
+            environments = await client.list_environments()
         except Exception as e:
-            self.log.exception(f"Error fetching runtime template list: {str(e)}")
+            self.log.exception(f"Error fetching composer environments: {str(e)}")
             self.finish({"error": str(e)})
+
+        response = []
+        for environment in environments:
+            env = environment.dict()
+            response.append(env)
+        self.finish(json.dumps(response))
