+++ dataproc_jupyter_plugin/controllers/dataproc.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Google LLC
+# Copyright 2024 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,12 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 import json
 from dataproc_jupyter_plugin import credentials
 from jupyter_server.base.handlers import APIHandler
 import tornado
-from dataproc_jupyter_plugin.services.clusterListService import ClusterListService
+from dataproc_jupyter_plugin.services import dataproc
 
 
 class ClusterListController(APIHandler):
@@ -26,12 +25,22 @@ class ClusterListController(APIHandler):
         try:
             page_token = self.get_argument("pageToken")
             page_size = self.get_argument("pageSize")
-            cluster = ClusterListService()
-            gcp_credentials = await credentials.get_cached()
-            cluster_list = cluster.list_clusters(
-                gcp_credentials, page_size, page_token, self.log
-            )
+            client = dataproc.Client(await credentials.get_cached(), self.log)
+            cluster_list = await client.list_clusters(page_size, page_token)
             self.finish(json.dumps(cluster_list))
         except Exception as e:
             self.log.exception(f"Error fetching cluster list")
+            self.finish({"error": str(e)})
+
+class RuntimeController(APIHandler):
+    @tornado.web.authenticated
+    async def get(self):
+        try:
+            page_token = self.get_argument("pageToken")
+            page_size = self.get_argument("pageSize")
+            client = dataproc.Client(await credentials.get_cached(), self.log)
+            runtime_list = await client.list_runtime(page_size, page_token)
+            self.finish(json.dumps(runtime_list))
+        except Exception as e:
+            self.log.exception(f"Error fetching runtime template list: {str(e)}")
             self.finish({"error": str(e)})
