+++ dataproc_jupyter_plugin/controllers/executor.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Google LLC
+# Copyright 2024 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,11 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
+import json
 from dataproc_jupyter_plugin import credentials
 from jupyter_server.base.handlers import APIHandler
 import tornado
 from dataproc_jupyter_plugin.services.executorService import ExecutorService
+from dataproc_jupyter_plugin.services import executor
 
 
 class ExecutorController(APIHandler):
@@ -24,9 +25,24 @@ class ExecutorController(APIHandler):
     async def post(self):
         try:
             input_data = self.get_json_body()
-            execute = ExecutorService()
-            gcp_credentials = await credentials.get_cached()
-            execute.execute(gcp_credentials, input_data, self.log)
+            client = executor.Client(await credentials.get_cached(), self.log)
+            await client.execute(input_data)
         except Exception as e:
             self.log.exception(f"Error creating dag schedule: {str(e)}")
+            self.finish({"error": str(e)})
+
+class DownloadOutputController(APIHandler):
+    @tornado.web.authenticated
+    async def get(self):
+        try:
+            client = executor.Client(await credentials.get_cached(), self.log)
+            bucket_name = self.get_argument("bucket_name")
+            dag_id = self.get_argument("dag_id")
+            dag_run_id = self.get_argument("dag_run_id")
+            download_status = client.download_dag_output(
+                bucket_name, dag_id, dag_run_id, self.log
+            )
+            self.finish(json.dumps({"status": download_status}))
+        except Exception as e:
+            self.log.exception(f"Error download output file")
             self.finish({"error": str(e)})
