+++ dataproc_jupyter_plugin/services/bigquery.py
@@ -16,7 +16,7 @@
 import requests
 
 from dataproc_jupyter_plugin import urls
-from dataproc_jupyter_plugin.commons.constants import CONTENT_TYPE
+from dataproc_jupyter_plugin.commons.constants import BIGQUERY_SERVICE_NAME, CLOUDRESOURCEMANAGER_SERVICE_NAME, CONTENT_TYPE, DATACATALOG_SERVICE_NAME
 
 
 class Client:
@@ -41,7 +41,7 @@ class Client:
 
     async def list_datasets(self, page_token, project_id):
         try:
-            bigquery_url = await urls.gcp_service_url("bigquery")
+            bigquery_url = await urls.gcp_service_url(BIGQUERY_SERVICE_NAME)
             api_endpoint = f"{bigquery_url}bigquery/v2/projects/{project_id}/datasets?pageToken={page_token}"
             response = requests.get(api_endpoint, headers=self.create_headers())
             if response.status_code == 200:
@@ -55,7 +55,7 @@ class Client:
 
     async def list_table(self, dataset_id, page_token, project_id):
         try:
-            bigquery_url = await urls.gcp_service_url("bigquery")
+            bigquery_url = await urls.gcp_service_url(BIGQUERY_SERVICE_NAME)
             api_endpoint = f"{bigquery_url}bigquery/v2/projects/{project_id}/datasets/{dataset_id}/tables?pageToken={page_token}"
             response = requests.get(api_endpoint, headers=self.create_headers())
             if response.status_code == 200:
@@ -69,7 +69,7 @@ class Client:
 
     async def list_dataset_info(self, dataset_id, project_id):
         try:
-            bigquery_url = await urls.gcp_service_url("bigquery")
+            bigquery_url = await urls.gcp_service_url(BIGQUERY_SERVICE_NAME)
             api_endpoint = f"{bigquery_url}bigquery/v2/projects/{project_id}/datasets/{dataset_id}"
             response = requests.get(api_endpoint, headers=self.create_headers())
             if response.status_code == 200:
@@ -83,7 +83,7 @@ class Client:
 
     async def list_table_info(self, dataset_id, table_id, project_id):
         try:
-            bigquery_url = await urls.gcp_service_url("bigquery")
+            bigquery_url = await urls.gcp_service_url(BIGQUERY_SERVICE_NAME)
             api_endpoint = f"{bigquery_url}bigquery/v2/projects/{project_id}/datasets/{dataset_id}/tables/{table_id}"
             response = requests.get(api_endpoint, headers=self.create_headers())
             if response.status_code == 200:
@@ -104,7 +104,7 @@ class Client:
         project_id,
     ):
         try:
-            bigquery_url = await urls.gcp_service_url("bigquery")
+            bigquery_url = await urls.gcp_service_url(BIGQUERY_SERVICE_NAME)
             api_endpoint = f"{bigquery_url}bigquery/v2/projects/{project_id}/datasets/{dataset_id}/tables/{table_id}/data?maxResults={max_results}&startIndex={start_index}"
             response = requests.get(api_endpoint, headers=self.create_headers())
             if response.status_code == 200:
@@ -118,7 +118,7 @@ class Client:
 
     async def bigquery_search(self, search_string, type, system, projects):
         try:
-            datacatalog_url = await urls.gcp_service_url("datacatalog")
+            datacatalog_url = await urls.gcp_service_url(DATACATALOG_SERVICE_NAME)
             api_endpoint = f"{datacatalog_url}v1/catalog:search"
             headers = {
                 "Content-Type": CONTENT_TYPE,
@@ -154,7 +154,7 @@ class Client:
 
     async def bigquery_projects(self, dataset_id, table_id):
         try:
-            cloudresourcemanager_url = await urls.gcp_service_url("cloudresourcemanager")
+            cloudresourcemanager_url = await urls.gcp_service_url(CLOUDRESOURCEMANAGER_SERVICE_NAME)
             api_endpoint = f"{cloudresourcemanager_url}v1/projects"
             response = requests.get(api_endpoint, headers=self.create_headers())
             if response.status_code == 200:
